aicloudops/logger/requirements.txt
pika==1.3.2
python-dotenv==1.0.1
SQLAlchemy==2.0.27
psycopg2==2.9.9
